Renamed multisignature to multisignatureUtils
